Divide index.spec.js into update.spec.js and wrapCursor.spec.js
